Extract common functions for dealing with private_envs json
